+++ httpgrpc/server/server.go
@@ -5,10 +5,8 @@
 package server
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"io"
 	"net"
 	"net/http"
 	"net/http/httptest"
@@ -27,13 +25,6 @@ import (
 	"github.com/grafana/dskit/middleware"
 )
 
-var (
-	// DoNotLogErrorHeaderKey is a header key used for marking non-loggable errors. More precisely, if an HTTP response
-	// has a status code 5xx, and contains a header with key DoNotLogErrorHeaderKey and any values, the generated error
-	// will be marked as non-loggable.
-	DoNotLogErrorHeaderKey = http.CanonicalHeaderKey("X-DoNotLogError")
-)
-
 // Server implements HTTPServer.  HTTPServer is a generated interface that gRPC
 // servers must implement.
 type Server struct {
@@ -47,18 +38,9 @@ func NewServer(handler http.Handler) *Server {
 	}
 }
 
-type nopCloser struct {
-	*bytes.Buffer
-}
-
-func (nopCloser) Close() error { return nil }
-
-// BytesBuffer returns the underlaying `bytes.buffer` used to build this io.ReadCloser.
-func (n nopCloser) BytesBuffer() *bytes.Buffer { return n.Buffer }
-
 // Handle implements HTTPServer.
 func (s Server) Handle(ctx context.Context, r *httpgrpc.HTTPRequest) (*httpgrpc.HTTPResponse, error) {
-	req, err := UnwrapHTTPRequest(ctx, r)
+	req, err := httpgrpc.ToHTTPRequest(ctx, r)
 	if err != nil {
 		return nil, err
 	}
@@ -68,12 +50,12 @@ func (s Server) Handle(ctx context.Context, r *httpgrpc.HTTPRequest) (*httpgrpc.
 	header := recorder.Header()
 	resp := &httpgrpc.HTTPResponse{
 		Code:    int32(recorder.Code),
-		Headers: fromHeader(header),
+		Headers: httpgrpc.FromHeader(header),
 		Body:    recorder.Body.Bytes(),
 	}
 	if recorder.Code/100 == 5 {
 		err := httpgrpc.ErrorFromHTTPResponse(resp)
-		if _, ok := header[DoNotLogErrorHeaderKey]; ok {
+		if _, ok := header[httpgrpc.DoNotLogErrorHeaderKey]; ok {
 			err = middleware.DoNotLogError{Err: err}
 		}
 		return nil, err
@@ -161,51 +143,6 @@ func NewClient(address string) (*Client, error) {
 	}, nil
 }
 
-// WrapHTTPRequest wraps an ordinary http.Request up into an httpgrpc.HTTPRequest
-func WrapHTTPRequest(r *http.Request) (*httpgrpc.HTTPRequest, error) {
-	body, err := io.ReadAll(r.Body)
-	if err != nil {
-		return nil, err
-	}
-	return &httpgrpc.HTTPRequest{
-		Method:  r.Method,
-		Url:     r.RequestURI,
-		Body:    body,
-		Headers: fromHeader(r.Header),
-	}, nil
-}
-
-// UnwrapHTTPRequest unwraps an ordinary http.Request from an httpgrpc.HTTPRequest
-func UnwrapHTTPRequest(ctx context.Context, r *httpgrpc.HTTPRequest) (*http.Request, error) {
-	req, err := http.NewRequest(r.Method, r.Url, nopCloser{Buffer: bytes.NewBuffer(r.Body)})
-	if err != nil {
-		return nil, err
-	}
-	toHeader(r.Headers, req.Header)
-	req = req.WithContext(ctx)
-	req.RequestURI = r.Url
-	req.ContentLength = int64(len(r.Body))
-	return req, nil
-}
-
-// WriteResponse converts an httpgrpc response to an HTTP one
-func WriteResponse(w http.ResponseWriter, resp *httpgrpc.HTTPResponse) error {
-	toHeader(resp.Headers, w.Header())
-	w.WriteHeader(int(resp.Code))
-	_, err := w.Write(resp.Body)
-	return err
-}
-
-// WriteError converts an httpgrpc error to an HTTP one
-func WriteError(w http.ResponseWriter, err error) {
-	resp, ok := httpgrpc.HTTPResponseFromError(err)
-	if ok {
-		_ = WriteResponse(w, resp)
-	} else {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-	}
-}
-
 // ServeHTTP implements http.Handler
 func (c *Client) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	if tracer := opentracing.GlobalTracer(); tracer != nil {
@@ -216,7 +153,7 @@ func (c *Client) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
-	req, err := WrapHTTPRequest(r)
+	req, err := httpgrpc.FromHTTPRequest(r)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
@@ -233,28 +170,8 @@ func (c *Client) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
-	if err := WriteResponse(w, resp); err != nil {
+	if err := httpgrpc.WriteResponse(w, resp); err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
-}
-
-func toHeader(hs []*httpgrpc.Header, header http.Header) {
-	for _, h := range hs {
-		header[h.Key] = h.Values
-	}
-}
-
-func fromHeader(hs http.Header) []*httpgrpc.Header {
-	result := make([]*httpgrpc.Header, 0, len(hs))
-	for k, vs := range hs {
-		if k == DoNotLogErrorHeaderKey {
-			continue
-		}
-		result = append(result, &httpgrpc.Header{
-			Key:    k,
-			Values: vs,
-		})
-	}
-	return result
 }
